+++ client/src/game/shapes/multiline.ts
@@ -21,7 +21,7 @@ export class MultiLine extends Shape {
         this.lineWidth = lineWidth || 3;
     }
 
-    get refPoint() {
+    get refPoint(): GlobalPoint {
         return this._refPoint;
     }
     set refPoint(point: GlobalPoint) {
@@ -30,17 +30,18 @@ export class MultiLine extends Shape {
         for (let i = 0; i < this._points.length; i++) this._points[i] = this._points[i].add(delta);
     }
 
-    asDict() {
+    asDict(): ServerMultiLine {
         return Object.assign(this.getBaseDict(), {
+            // eslint-disable-next-line @typescript-eslint/camelcase
             line_width: this.lineWidth,
             points: this._points.map(p => ({ x: p.x, y: p.y })),
         });
     }
-    fromDict(data: ServerMultiLine) {
+    fromDict(data: ServerMultiLine): void {
         super.fromDict(data);
         this._points = data.points.map(p => new GlobalPoint(p.x, p.y));
     }
-    get points() {
+    get points(): number[][] {
         return this._points.map(point => [point.x, point.y]);
     }
     getBoundingBox(): BoundingRect {
@@ -56,7 +57,7 @@ export class MultiLine extends Shape {
         }
         return new BoundingRect(new GlobalPoint(minx, miny), maxx - minx, maxy - miny);
     }
-    draw(ctx: CanvasRenderingContext2D) {
+    draw(ctx: CanvasRenderingContext2D): void {
         super.draw(ctx);
         ctx.beginPath();
         ctx.lineCap = "round";
@@ -74,7 +75,7 @@ export class MultiLine extends Shape {
 
     center(): GlobalPoint;
     center(centerPoint: GlobalPoint): void;
-    center(centerPoint?: GlobalPoint): GlobalPoint | void {
+    center(_centerPoint?: GlobalPoint): GlobalPoint | void {
         return this.getBoundingBox().center();
     }
     visibleInCanvas(canvas: HTMLCanvasElement): boolean {