+++ client/src/game/shapes/rect.ts
@@ -3,16 +3,17 @@ import { BaseRect } from "@/game/shapes/baserect";
 import { gameStore } from "@/game/store";
 import { g2l } from "@/game/units";
 import { getFogColour } from "@/game/utils";
+import { ServerShape } from "../comm/types/shapes";
 
 export class Rect extends BaseRect {
     type = "rect";
     constructor(topleft: GlobalPoint, w: number, h: number, fillColour?: string, strokeColour?: string, uuid?: string) {
         super(topleft, w, h, fillColour, strokeColour, uuid);
     }
-    asDict() {
+    asDict(): ServerShape {
         return super.getBaseDict();
     }
-    draw(ctx: CanvasRenderingContext2D) {
+    draw(ctx: CanvasRenderingContext2D): void {
         super.draw(ctx);
         if (this.fillColour === "fog") ctx.fillStyle = getFogColour();
         else ctx.fillStyle = this.fillColour;