+++ client/src/game/ui/tools/pan.ts
@@ -1,3 +1,5 @@
+import Component from "vue-class-component";
+
 import Tool from "@/game/ui/tools/tool.vue";
 
 import { sendClientOptions } from "@/game/api/utils";
@@ -5,7 +7,6 @@ import { LocalPoint } from "@/game/geom";
 import { layerManager } from "@/game/layers/manager";
 import { gameStore } from "@/game/store";
 import { getMouse } from "@/game/utils";
-import Component from "vue-class-component";
 
 @Component
 export default class PanTool extends Tool {
@@ -13,11 +14,11 @@ export default class PanTool extends Tool {
     panStart = new LocalPoint(0, 0);
     active = false;
 
-    onMouseDown(event: MouseEvent) {
+    onMouseDown(event: MouseEvent): void {
         this.panStart = getMouse(event);
         this.active = true;
     }
-    onMouseMove(event: MouseEvent) {
+    onMouseMove(event: MouseEvent): void {
         if (!this.active) return;
         const mouse = getMouse(event);
         const distance = mouse.subtract(this.panStart).multiply(1 / gameStore.zoomFactor);
@@ -26,8 +27,8 @@ export default class PanTool extends Tool {
         this.panStart = mouse;
         layerManager.invalidate();
     }
-    onMouseUp(event: MouseEvent) {
+    onMouseUp(_event: MouseEvent): void {
         this.active = false;
-        sendClientOptions();
+        sendClientOptions(gameStore.locationOptions);
     }
 }