+++ client/src/invitation/invitation.ts
@@ -5,14 +5,14 @@ import { Route } from "vue-router";
 
 @Component({ template: "" })
 export default class Initiative extends Vue {
-    beforeRouteEnter(to: Route, from: Route, next: ({}) => {}) {
+    beforeRouteEnter(to: Route, from: Route, next: (nextInfo: { path: string }) => {}): void {
         Axios.post("/api/invite", {
             code: to.params.code,
         })
             .then((response: AxiosResponse) => {
                 next({ path: response.data.sessionUrl });
             })
-            .catch((error: AxiosError) => {
+            .catch((_error: AxiosError) => {
                 console.error("Invitation code could not be redeemed");
                 next({ path: "/dashboard" });
             });