+++ client/src/shims/shims-tsx.d.ts
@@ -1,13 +1,13 @@
-import Vue, { VNode } from 'vue';
+import Vue, { VNode } from "vue";
 
 declare global {
-  namespace JSX {
-    // tslint:disable no-empty-interface
-    interface Element extends VNode {}
-    // tslint:disable no-empty-interface
-    interface ElementClass extends Vue {}
-    interface IntrinsicElements {
-      [elem: string]: any;
+    namespace JSX {
+        // eslint-disable-next-line @typescript-eslint/no-empty-interface
+        interface Element extends VNode {}
+        // eslint-disable-next-line @typescript-eslint/no-empty-interface
+        interface ElementClass extends Vue {}
+        interface IntrinsicElements {
+            [elem: string]: any;
+        }
     }
-  }
 }